requirements.txt
django
django-autoslug
django-debug-toolbar
django-crispy-forms
django-filter
django-grappelli
django-object-actions
django-pdb
djangorestframework
docker
docker-compose
factory-boy
ipython
isort
Pillow
psycopg2
pytest
pytest-cov
pytest-django